파일 경로, type 컨벤션 수정
